+++ CHANGES.rst
@@ -5,7 +5,7 @@ Change log for sw.allotmentclub
 4.9.3 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Use sentry in frontend for error reporting.
 
 
 4.9.2 (2019-10-29)